+++ src/App/index.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import { colors } from 'data';
+import Autocomplete from 'App/Autocomplete';
+
+// import styles from './App.module.scss';
+
+class App extends Component {
+  render() {
+    return (
+      <Autocomplete suggestions={colors()}/>
+    );
+  }
+}
+
+export default App;
